consolidate usePlacesClient into react-shared lib
