+++ libs/react/shared/src/lib/hooks/index.ts
@@ -1,3 +1,5 @@
 export { useAllauthLogin } from './useAllauthLogin';
 export type { FetchFn, LoginStep } from './useAllauthLogin';
 export { useDebounce } from './useDebounce';
+export { usePlacesClient } from './usePlacesClient';
+export type { TPlacesClient } from './usePlacesClient';
