+++ libs/react/shared/src/lib/hooks/usePlacesClient.ts
@@ -7,11 +7,30 @@ import {
 import { useMapsLibrary } from '@vis.gl/react-google-maps';
 import { useCallback, useMemo } from 'react';
 
+export type TPlacesClient = {
+  autocomplete: (
+    query: string,
+    options?: {
+      boundsCenter?: { latitude: number; longitude: number };
+      boundsRadiusMiles?: number;
+      includedRegionCodes?: string[];
+    },
+  ) => Promise<TPlacePrediction[]>;
+  getDetails: (
+    placeId: string,
+    options?: { fields?: string },
+  ) => Promise<TPlaceDetails>;
+  reverseGeocode: (
+    latitude: number,
+    longitude: number,
+  ) => Promise<TReverseGeocodeResult>;
+};
+
 /**
  * Places client backed by the Google Maps JS SDK libraries
  * loaded via `<APIProvider>` from `@vis.gl/react-google-maps`.
  */
-export function usePlacesClient() {
+export function usePlacesClient(): TPlacesClient {
   const placesLib = useMapsLibrary('places');
   const geocodingLib = useMapsLibrary('geocoding');
 
@@ -22,7 +41,7 @@ export function usePlacesClient() {
         boundsCenter?: { latitude: number; longitude: number };
         boundsRadiusMiles?: number;
         includedRegionCodes?: string[];
-      }
+      },
     ): Promise<TPlacePrediction[]> => {
       if (!placesLib || query.length < 3) return [];
 
@@ -46,7 +65,7 @@ export function usePlacesClient() {
 
       const { suggestions } =
         await google.maps.places.AutocompleteSuggestion.fetchAutocompleteSuggestions(
-          request
+          request,
         );
 
       return suggestions
@@ -64,13 +83,13 @@ export function usePlacesClient() {
           };
         });
     },
-    [placesLib]
+    [placesLib],
   );
 
   const getDetails = useCallback(
     async (
       placeId: string,
-      options?: { fields?: string }
+      options?: { fields?: string },
     ): Promise<TPlaceDetails> => {
       if (!placesLib) throw new Error('Places library not loaded');
 
@@ -108,17 +127,17 @@ export function usePlacesClient() {
             longText: c.longText || '',
             shortText: c.shortText || '',
             types: c.types,
-          })
+          }),
         ),
       };
     },
-    [placesLib]
+    [placesLib],
   );
 
   const reverseGeocode = useCallback(
     async (
       latitude: number,
-      longitude: number
+      longitude: number,
     ): Promise<TReverseGeocodeResult> => {
       if (!geocodingLib) throw new Error('Geocoding library not loaded');
 
@@ -142,15 +161,15 @@ export function usePlacesClient() {
             longText: c.long_name,
             shortText: c.short_name,
             types: c.types,
-          })
+          }),
         ),
       };
     },
-    [geocodingLib]
+    [geocodingLib],
   );
 
   return useMemo(
     () => ({ autocomplete, getDetails, reverseGeocode }),
-    [autocomplete, getDetails, reverseGeocode]
+    [autocomplete, getDetails, reverseGeocode],
   );
 }
