--- libs/react/shelter/src/lib/places/index.ts
@@ -1,1 +1,0 @@
-export { usePlacesClient } from './usePlacesClient';
